refactor code per review comments
